+++ Assets/04.Scenes/LoginScene.unity
@@ -6417,9 +6417,9 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 2bea457ffadcc4e42a58918bce360d68, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  _maxOnlineUser: 0
   _loginButton: {fileID: 1121620500}
   _signupButton: {fileID: 1907427290}
+  _maxOnlineUser: 1
 --- !u!114 &713239657
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -14231,9 +14231,9 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 2bea457ffadcc4e42a58918bce360d68, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  _maxOnlineUser: 5
   _loginButton: {fileID: 2128643558}
   _signupButton: {fileID: 896720086}
+  _maxOnlineUser: 2
 --- !u!114 &1628973895
 MonoBehaviour:
   m_ObjectHideFlags: 0
