+++ Assets/07.CYH_Folder/Scripts/LoginPanel.cs
@@ -14,7 +14,6 @@ public class LoginPanel : UIBase
 
     public Action OnClickSignup { get; set; }
     public Action OnClickLogin { get; set; }
-    public Action MaxUser { get; set; }
 
     private void Start()
     {
@@ -26,48 +25,5 @@ public class LoginPanel : UIBase
         {
             OnClickSignup?.Invoke();
         });
-    }
-
-    private async void OnEnable()
-    {
-        CurrentUserCount();
-    }
-
-    /// <summary>
-    /// 현재 접속 인원을 확인하는 메서드
-    /// 최대 인원을 초과한 경우 안내 팝업 호출
-    /// </summary>
-    public async void CurrentUserCount()
-    {
-        if (!await CurrentOnlineUserCount(20))
-        {
-            PopupManager.Instance.ShowOKPopup($"접속 인원이 초과되어 입장할 수 없습니다.\n(Max: 20명)", "OK", () => PopupManager.Instance.HidePopup());
-
-            MaxUser?.Invoke();
-        }
-    }
-
-    /// <summary>
-    /// 현재 접속 중인 유저 수를 확인하고 최대 접속 인원 초과 여부를 반환하는 메서드
-    /// </summary>
-    /// <param name="maxOnlineUser">최대 접속 가능 인원 수 (기본값: 20)</param>
-    /// <returns>접속 가능 여부 (true: 가능, false: 인원 초과)</returns>
-    public async Task<bool> CurrentOnlineUserCount(int maxOnlineUser = 20)
-    {
-        DatabaseReference userDataRef = CYH_FirebaseManager.DataReference.Child("UserData");
-        DataSnapshot snapshot = await userDataRef.GetValueAsync();
-
-        int onlineCount = 0;
-        foreach (var user in snapshot.Children)
-        {
-            var isOnlineValue = user.Child("IsOnline").Value;
-            if (isOnlineValue != null && isOnlineValue.ToString() == "True")
-            {
-                onlineCount++;
-            }
-        }
-
-        Debug.Log($"현재 접속 중인 인원 : {onlineCount}");
-        return onlineCount < maxOnlineUser;
     }
 }
