+++ Assets/07.CYH_Folder/Scripts/LoginSceneUIController.cs
@@ -77,14 +77,6 @@ public class LoginSceneUIManager : MonoBehaviour
                 // 로그인 버튼
                 loginPanel.OnClickLogin = () => ShowUI(LoginUIType.LoginOptionPanel);
                 loginPanel.OnClickSignup = () => ShowUI(LoginUIType.SignUpPanel);
-                //loginPanel.OnClickSocialLogin = () => ShowUI(LoginUIType.LoginOptionPanel);
-                
-                // 접속 정원 초과
-                loginPanel.MaxUser = () =>
-                {
-                    HideUI(LoginUIType.LoginPanel);
-                    ShowUI(LoginUIType.MainPanel);
-                };
             }
 
             else if (ui is LoginOptionPanel loginOptionPanel)
